Download new events only if a push notification was received
This change introduces 2 synchronization strategies: light and full.
A light synchronization does not fetch new events from the remote
server, whereas a full synchronization does it.

A full synchronization is triggered when a C2DM push notification is
received, or when an account is selected.

This should helps both client and server to save resources when no
new remote events are available.