Prevent synchronization from running if the device is already synchronized
In some cases, several C2DM registration ids could be linked to the
same device. If the user account is updated, the same device may
receive several requests for synchronization.

When new data is sent to the server, a synchronization request is also
sent with a random token. The server pushes this synchronization request
to every user devices: if the received token is equal to the last sent
token, the synchronization is not started.